% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSelection.R
\name{ApplyCountThreshold}
\alias{ApplyCountThreshold}
\title{Remove rare and ubiquitous genes}
\usage{
ApplyCountThreshold(M, countL = 1, countU = 3, X = 3)
}
\arguments{
\item{M}{a matrix of expression values for each gene (rows) and cell (columns)}

\item{countL}{the definition of "expressed" raw counts for lower threshold}

\item{countU}{the definition of "expressed" raw counts for upper threshold}

\item{X}{the threshold percentage: lower = X\%, upper = 100-X\%}
}
\value{
indices of genes to keep
}
\description{
This is based on Kiselev et al 2017 (SC3).  We remove genes where we observe at least countL copies in less than X% of cells or at least countU copies in more than 100 - X% of cells.  Adjust countL down and countU up in order to get more permissive filters.
}
