% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindRootCluster.R
\name{AvgFeatureExpression}
\alias{AvgFeatureExpression}
\title{Find average cluster feature expression}
\usage{
AvgFeatureExpression(cluster_labels, M, feature, feature_list)
}
\arguments{
\item{cluster_labels}{the cluster label for each cell}

\item{M}{a matrix of expression values for each cell}

\item{feature}{a feature id, usually a gene name}

\item{feature_list}{the names of the features represented by rows of M}
}
\value{
a vector of feature expression for each cluster
}
\description{
Average the expression of a feature and return the expression vector.
}
