% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signaling.R
\name{ClusterAvg}
\alias{ClusterAvg}
\title{Create a heatmap with signaling markers over clusters}
\usage{
ClusterAvg(M, gene_names, cell_order, cell_labels, gene_list)
}
\arguments{
\item{M}{a matrix of genes x cells where each entry is normalized expression}

\item{gene_names}{the names of the genes in the rows of M}

\item{cell_order}{permutation of the labels of P, cells ordered by cluster}

\item{cell_labels}{cluster labels of the permuted cells}

\item{gene_list}{filter by a list of gene names}
}
\value{
a matrix of avg expression values
}
\description{
Create a heatmap with signaling markers over clusters.
}
