% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{ClusterCells}
\alias{ClusterCells}
\title{Cluster the Cells}
\usage{
ClusterCells(similarityMatrix = NULL, n_clusters = NULL, n_comp = 3,
  ...)
}
\arguments{
\item{similarityMatrix}{a symmetric nonnegative similarity matrix}

\item{n_clusters}{default is NULL, ensemble method will be called}

\item{n_comp}{the number of similarity components to use for ensemble clustering}

\item{...}{additional parameters for NMF}
}
\value{
a list containing 
    \item{H}{the cluster weight matrix, the factorization of the similarity matrix}
    \item{labels}{the cluster labels}
}
\description{
Use NMF to cluster the cells.  Currently, the Lee and Seung algorithm is implemented via a call to renozao/NMF.  If cluster-number is inferred then output relevent diagnostics (eg the eigenvalue spaceing of the ensemble method).
}
