% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signaling.R
\name{ClusterSig}
\alias{ClusterSig}
\title{Average cluster to cluster gene expression}
\usage{
ClusterSig(P, cell_order, cell_labels)
}
\arguments{
\item{P}{signaling probabilities cells x cells}

\item{cell_order}{permutation of the labels of P, cells ordered by cluster}

\item{cell_labels}{cluster labels of the permuted cells}
}
\description{
Average cluster to cluster gene expression
}
