% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signaling.R
\name{ClusterSig_single}
\alias{ClusterSig_single}
\title{Average cluster to cluster signaling.}
\usage{
ClusterSig_single(Pcell, cluster_labels, normalize_rows = FALSE)
}
\arguments{
\item{Pcell}{signaling probabilities cells x cells}

\item{cluster_labels}{labels of cells 1:n}

\item{normalize_rows}{normalize the rows of the output matrix so that the rows sum to 1, default true}
}
\value{
a matrix of cluster to cluster signaling
}
\description{
Average cluster to cluster signaling.
}
