% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimilarityM.R
\name{computM}
\alias{computM}
\title{Perform ADMM}
\usage{
computM(D, X, lambda)
}
\arguments{
\item{D}{the unweighted KNN adjacency matrix}

\item{X}{normalized sample vectors}

\item{lambda}{the balance term between the rank of Z and the error}
}
\value{
a list containing the low rank approximation of X and manifold learning error
}
\description{
Computing cell-to-cell similarity matrix by solving the following
optimization problem via ADMM
}
\details{
min_{Z,E}  ||Z||_* + lambda ||E||_{2,1}
  s.t.       X = XZ + E;
             Z'1 = 1;
              Z_{i,j} = 0 for (i,j)\in Omega
}
