% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{FeatureScatterPlot}
\alias{FeatureScatterPlot}
\title{Produce a scatter plot of the cells on selected 2-dim ebedding colored by feature}
\usage{
FeatureScatterPlot(flat_embedding, feature, outputdir = NULL,
  outputfile = NULL)
}
\arguments{
\item{flat_embedding}{a low dim embedding of cells}

\item{feature}{a scalar representation of feature}

\item{outputdir}{the output directory, relative to getwd()}

\item{outputfile}{the output file}

\item{title}{the title of the plot}

\item{subtitle}{the subtitle of the plot}
}
\value{
a ggplot2 object
}
\description{
If an output dir and filename are provided, a plot will
be saved
}
