% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindRootCell.R
\name{FindRootCell}
\alias{FindRootCell}
\title{Infer the root cell}
\usage{
FindRootCell(use_flat_dist = TRUE, use_avg_length = TRUE,
  graph_cluster = NULL, dist_graph = NULL, dist_flat,
  cluster_labels = NULL, root_cluster = NULL)
}
\arguments{
\item{use_flat_dist}{use max path length heuristic on flat embedding}

\item{use_avg_length}{use the avg path length from putative root cell to cluster k, otherwise use avg cell index}

\item{graph_cluster}{the weighted adjacency matrix for a graph of clusters}

\item{dist_graph}{a distance matrix of cells embedded in a graph}

\item{dist_flat}{the manifold embedding of the cells}

\item{cluster_labels}{the cluster label for each cell}

\item{root_cluster}{the id of the root cluster on the cluster-cluster graph}
}
\value{
integer index of the root cell
}
\description{
Infer the root cell
}
\examples{
root_cell <- FindRootCell(use_flat_dist = TRUE, <some_pca_embeddings>)

Use either the primary manifold embedding of cell similarity to find the root cell by cluster-path-length/pseudotime correlation
 or use the flattened representation of this embedding to find the root cell by maximum separation heuristic
}
