% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindRootCluster.R
\name{FindRootCluster}
\alias{FindRootCluster}
\title{Find the root cluster}
\usage{
FindRootCluster(cluster_labels, flat_embedding, dist_graph, dist_flat,
  reverse = FALSE)
}
\arguments{
\item{cluster_labels}{the cluster label for each cell}

\item{flat_embedding}{rows are cells and columns are coordinates in n-col space}

\item{dist_graph}{distances on a connected graph}

\item{dist_flat}{flat distance matrix for all cells}

\item{reverse}{a boolean variable whether to take the root cluster based on minimum dispersion on the flat embedding}
}
\value{
a list containing:
    \item{cluster_adj_matrix}{a weighted upper-triangular adjacency matrix for the clusters based on avg pseudotime between their cells}
    \item{graph_cluster}{an igraph object on the clusters, with a minimum spanning tree}
    \item{root_cluster}{index of the root cluster}
    \item{cluster_mst}{an igraph mst object}
}
\description{
Given a weighted adjacency matrix, generate a cluser-to-cluster graph based on a similarity-matrix of connected cells.  Generate the minimum spanning tree on the cluster-to-cluster graph.  Use one of the max dist clusters from this graph, unless:
    The cluster is given, then proceed to find the root cell
    The root cell is given, then use its cluster
}
