% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindRootCell.R
\name{GetCellMetric}
\alias{GetCellMetric}
\title{Get metric of pseudotime}
\usage{
GetCellMetric(cell_order_by, cell_id, cluster_id, cluster_labels,
  dist_graph)
}
\arguments{
\item{cell_order_by}{cell-wise pseudotime parameter for measuring cluster/cell rank correlation}

\item{cell_id}{the cell to measure from}

\item{cluster_id}{the cluster of the cell}

\item{cluster_labels}{the cluster label for each cell}

\item{dist_graph}{a distance matrix of cells embedded in a graph}
}
\value{
the mean distance (by given metric) from the given cell
}
\description{
Measure pseudotime from the given cell to its cluster partners.
}
