% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindRootCell.R
\name{GetCorrelation}
\alias{GetCorrelation}
\title{Get rank correlation}
\usage{
GetCorrelation(cluster_order_by, cell_metric, graph_cluster_mst,
  root_cluster)
}
\arguments{
\item{cluster_order_by}{cluster rank parameter for measuring cluster/cell rank correlation}

\item{cell_metric}{either distance or index}

\item{graph_cluster_mst}{igraph object representing an mst on the cluster-cluster graph}

\item{root_cluster}{the id of the root cluster on the cluster-cluster graph}
}
\value{
tau score of rank correlation
}
\description{
Get rank correlation for pseudotime analysis
}
