% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeFunctions.R
\name{GetDGFromTable}
\alias{GetDGFromTable}
\title{Get a weighted, directed graph from a table of edges and weights}
\usage{
GetDGFromTable(directed_edge_table)
}
\arguments{
\item{directed_edge_table}{a numeric matrix whose rows are edges, col 1 is v1, col2 is v2, col3 is weight}
}
\value{
a directed igraph object
}
\description{
Get a weighted, directed graph from a table of edges and weights
}
