% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeFunctions.R
\name{GetDominatorTree}
\alias{GetDominatorTree}
\title{Get a directed graph from a predecessor vector}
\usage{
GetDominatorTree(predecessors)
}
\arguments{
\item{predecessors}{the tree to be directionalized}
}
\value{
a directed igraph object
}
\description{
Get a directed graph from a predecessor vector
}
