% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterNumber.R
\name{GetEnsemble}
\alias{GetEnsemble}
\title{Produce a truncated ensemble consensus matrix}
\usage{
GetEnsemble(data, tol, n_prcs = 3, tau, range = 1:20,
  method = "kmeans")
}
\arguments{
\item{data}{a symmetric nonnegative similarity matrix}

\item{tol}{cutoff for lambda zero}

\item{tau}{the drop tolerance, controlling the sparsification
(uncoupling) of the consensus matrix}

\item{range}{a vector specifying the min and max
number of clusters to iteratively test when building the
consensus matrix}

\item{method}{the clustering method for building consensus clusters}

\item{prcs_dim}{the number of pcs to use for clustering method}
}
\value{
a truncated ensemble consensus matrix
}
\description{
Produce a truncated ensemble consensus matrix
}
