% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeFunctions.R
\name{GetPredecessors}
\alias{GetPredecessors}
\title{Get the predecessor vector for a dominator tree encoded as an undirected tree and root}
\usage{
GetPredecessors(minspantree, root)
}
\arguments{
\item{minspantree}{the tree to be directionalized}

\item{root}{the id of the root}
}
\value{
a vector[] of node ids, where vector[i] is the predecessor of node i
}
\description{
Get the predecessor vector for a dominator tree encoded as an undirected tree and root
}
