% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE67602_Joost}
\alias{GSE67602_Joost}
\title{Skin Cell scRNA-seq Data from Joost et al}
\format{A list with three items:
\describe{
  \item{annotation}{the type of cell (Basal IFE-DI IFE-DII IFE-KI IFE-KII)}
  \item{data}{values are mRNA counts, row names are genes, column names are cells}
}}
\usage{
GSE67602_Joost
}
\description{
A dataset containing sequencing on over 20k genes across 719 annotated cells
}
\keyword{datasets}
