% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signaling.R
\name{ImportPathway}
\alias{ImportPathway}
\title{Process siganling pathway data}
\usage{
ImportPathway(lig_table = NULL, rec_table = NULL, lig_table_path,
  rec_table_path = NULL, data, gene_names = NULL)
}
\arguments{
\item{lig_table}{a data frame with "ligand" in column 1 and "receptor" in column 2}

\item{rec_table}{a data frame with "receptor" in column 1, "target" in column 2, and up/down "direction"}

\item{lig_table_path}{a tsv/csv file with "ligand" in column 1 and "receptor" in column 2}

\item{rec_table_path}{a tsv/csv file with "receptor" in column 1, "target" in column 2, and up/down "direction" in column 3}

\item{data}{a matrix of the data, with gene names as row names (required if no gene list provided)}

\item{gene_names}{an optional list of gene names (required if rownames of data are null)}
}
\value{
a list containing:
    \item{lignames}{vector of all ligands}
    \item{recnames}{vector of all receptors}
    \item{targets}{vector of all targets}
    \item{pathway}{a table with combined pathway information}
    \item{pathway_removed}{a table with markers excluded}
    \item{removed}{a table with the excluded markers}
}
\description{
Given two tsv tables, one including ligand-receptor pairs, and the other including receptor-target pairs, and a data set, inner-join the two tables and make sure no ligands, receptors, or targets are missing from the data.  This function returns a summary table of all requested markers.  If the data is a subset of some larger study, genes may be reported (a column exists in the data matrix) for which zero copies were observed.  In this case, "dropout in subset" applies.  There are also known gene symbols which were not measured in the study (no corresponding row exists in the data matrix).  In this case, "missing from all subsets" applies.
}
