% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitSVD.R
\name{InitSVD}
\alias{InitSVD}
\title{Initialize non-negative factorization of the similarity matrix}
\usage{
InitSVD(A, k)
}
\arguments{
\item{A}{The similarity matrix.}

\item{k}{The rank of the output.}
}
\value{
\code{W} and \code{H} such that \code{A} = \code{W} * \code{D} * \code{H}.
}
\description{
Initialize non-negative factorization of the similarity matrix
}
