% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{LoadData}
\alias{LoadData}
\title{Load sequencing data and annotations}
\usage{
LoadData(d_f, gene_f, cell_f, annotation_f)
}
\arguments{
\item{d_f}{path to the file containing counts}

\item{gene_f}{path the file containing unique gene ids}

\item{cell_f}{path the file containing unique cell ids}

\item{annotation_f}{path to the file containing labels, e.g. tissue type of the cells}
}
\value{
a list containing:
    \item{cell_names}{cell ids}
    \item{gene_names}{eg gene symbols}
    \item{data}{counts matrix}
    \item{annotation}{cell labels}
}
\description{
A matrix of counts with rows as genes and columns as cells must be provided.  Gene and cell ids are provided in separate text files.  For labeled data, the vector must be provided as a separate file.  The counts are loaded as a sparse matrix, the other values as character vectors.
}
