% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSelection.R
\name{LogNormalize}
\alias{LogNormalize}
\title{Scale and center a data matrix}
\usage{
LogNormalize(M, scale = 10000, normalize = TRUE)
}
\arguments{
\item{M}{a matrix of expression values for each gene (rows) and cell (columns)}

\item{scale}{the scaling factor}

\item{normalize}{a boolean whether to normalize the data}
}
\value{
a log normalized matrix of data
}
\description{
This function normalizes the data across cells, scales it and log transforms it
}
