% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{MarkerHeatmap}
\alias{MarkerHeatmap}
\title{Get the marker genes for each cluster}
\usage{
MarkerHeatmap(counts_data, cluster_labels, markerTable, range = c(-3, 3),
  n_markers = 0)
}
\arguments{
\item{counts_data}{a matrix of expression values for each cell (rows) and gene (columns)}

\item{cluster_labels}{a vector of cluster labels}

\item{range}{the interval over which centered, normalized expression is displayed. expression values below range[1] are increased to range[1].  values above range[2] are decreased to range[2].}

\item{n_markers}{number of marker genes per cluster to retrieve (0 = all)}

\item{markerTabele}{a matrix of expression values for each cell (rows) and gene (columns)}
}
\value{
a table of marker genes
}
\description{
Get the marker genes for each cluster
}
