% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepresentationMap.R
\name{MembershipList}
\alias{MembershipList}
\title{Get a list of membership vectors}
\usage{
MembershipList(membership_vector)
}
\arguments{
\item{membership_vector}{the component membership of cells}
}
\value{
where list[[i]] = c(cells in component i)
}
\description{
Get a list of membership vectors to facilitate component-joining
}
