% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signaling.R
\name{NormalizeGene}
\alias{NormalizeGene}
\title{Normalize signal across cells}
\usage{
NormalizeGene(M)
}
\arguments{
\item{M}{a matrix of expression values for each cell (columns) and gene (rows)}
}
\value{
a normalized vector of expression values
}
\description{
Normalize signal across cells
}
