% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signaling.R
\name{NormalizeSubset}
\alias{NormalizeSubset}
\title{Normalize a Subset of the data}
\usage{
NormalizeSubset(M, ids)
}
\arguments{
\item{M}{a matrix of expression values for each cell (columns) and gene (rows)}

\item{ind}{a vector of row indexes corresponding to genes}
}
\value{
a normalized matrix of expression values
}
\description{
Normalize a Subset of the data.
}
