% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signaling.R
\name{PenaltyCoeff}
\alias{PenaltyCoeff}
\title{Generate the divergence penalty coefficient}
\usage{
PenaltyCoeff(alpha = alpha, const = const, n_cells = n_cells)
}
\arguments{
\item{alpha}{a matrix of normalized LR expression values for each cell (columns) and gene (rows)}

\item{const}{either the normalized repression targets or the normalized expression targets}
}
\description{
This will approach 1 as the difference between alpha and \code{const} approaches 0
}
