% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{PlotClusterSig}
\alias{PlotClusterSig}
\title{Produce a plot of a cluster-cluster signaling network}
\usage{
PlotClusterSig(data)
}
\arguments{
\item{data}{a matrix P where p_i,j = the probability that cluster i signals to cluster j}
}
\value{
nothing
}
\description{
Produce a plot of a cluster-cluster signaling network
}
