% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{PlotLineage}
\alias{PlotLineage}
\title{Plot the pseudotime ordering of clusters}
\usage{
PlotLineage(network, d_thickness = c(1, 3))
}
\arguments{
\item{network}{an igraph network with weighted edges corresponding to the pseudotime distance}

\item{d_thickness}{controls edge width.  c(1, m) => m/distance from L2R2, c(0,m) => m.}
}
\value{
nothing
}
\description{
Plot the pseudotime ordering of clusters
}
