% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{PlotMatlabDtree}
\alias{PlotMatlabDtree}
\title{Produce a plot of matlab DTree data and return the object}
\usage{
PlotMatlabDtree(edge_table, predecessors, outputdir = NULL, outputfile = NULL)
}
\arguments{
\item{edge_table}{a numeric matrix whose rows are directed edges of a tree:
col 1 is v1, col2 is v2, col3 is weight}

\item{predecessors}{a vector of tree predecessors such that pred[i] = the predecessor of i}

\item{outputdir}{the output directory, relative to getwd()}

\item{outputfile}{the output file}
}
\value{
an igraph representation of the tree
}
\description{
If an output dir and filename are provided, a plot will be saved, otherwise the function will just return the graph.  This is used for internal analysis to compare with matlab implementation.
}
