% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotP_functions.R
\name{plotSig}
\alias{plotSig}
\title{Plot the signaling between individual cells}
\usage{
plotSig(P, cluster_label, threshold, sigonly = FALSE,
  plotclusternames = TRUE, subsample, nsample, outfile = "Psig.pdf",
  plottitle)
}
\arguments{
\item{P}{probability of signal between cell i and cell j}

\item{cluster_label}{the ordered vector of cluster labels for each cell}

\item{threshold}{the threshold for p_ij whether to report signal}

\item{sigonly}{default is false}

\item{plotclusternames}{boolean whether to plot the names}

\item{subsample}{boolean whether to subsample}

\item{nsample}{number of samples}

\item{plottitle}{the main title}
}
\value{
nothing
}
\description{
Plot the signaling between individual cells
}
