% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{PlotTopN}
\alias{PlotTopN}
\title{Heatmap of the top n markers}
\usage{
PlotTopN(data, cluster_labels, markers, n_features,
  y_lsize = ((length(unique(cluster_labels)) * n_features)/120) * 5,
  y_tsize = ((length(unique(cluster_labels)) * n_features)/120) * 15,
  x_lsize = ((length(unique(cluster_labels)) * n_features)/120) * 14,
  x_tsize = ((length(unique(cluster_labels)) * n_features)/120) * 15,
  ...)
}
\arguments{
\item{data}{expression data with genes x cells}

\item{cluster_labels}{a vector of cluster labels corresponding to the columns in the data matrix}

\item{markers}{a table of markers}

\item{n_features}{the top n features to plot}

\item{y_lsize}{y axis label size}

\item{y_tsize}{y axis title size}

\item{x_lsize}{x axis label size}

\item{x_tsize}{x axis title size}

\item{...}{extra args to heatmap.2}

\item{gene_names}{a vector of symbolic gene names corresponding to the rows in the data matrix}
}
\value{
a ggplot2 plot object
}
\description{
Plot the heatmap of the top n inferred markers for each cluster.
}
