% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{PlotTopN_Grid}
\alias{PlotTopN_Grid}
\title{Factor Grid heatmap of the top n markers}
\usage{
PlotTopN_Grid(
  data,
  cluster_labels,
  markers,
  n_features,
  y_lsize = ((length(unique(cluster_labels)) * n_features)/120) * 5,
  y_tsize = ((length(unique(cluster_labels)) * n_features)/120) * 15,
  x_lsize = ((length(unique(cluster_labels)) * n_features)/120) * 14,
  x_tsize = ((length(unique(cluster_labels)) * n_features)/120) * 15,
  use_z = TRUE,
  spacing = 0.1
)
}
\arguments{
\item{data}{expression data with genes x cells}

\item{cluster_labels}{a vector of cluster labels corresponding to the columns in the data matrix}

\item{markers}{a table of markers}

\item{n_features}{the top n features to plot}

\item{y_lsize}{y axis label size}

\item{y_tsize}{y axis title size}

\item{x_lsize}{x axis label size}

\item{x_tsize}{x axis title size}

\item{use_z}{use z-score per gene instead of raw expression}

\item{spacing}{the number of lines between grids}
}
\value{
as heatmap
}
\description{
Plot the heatmap of the top n inferred markers for each cluster.
}
