% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeFunctions.R
\name{ProcessMatlabDTree}
\alias{ProcessMatlabDTree}
\title{Convert a matlab edge table and predecessor list into a directed weighted edge table}
\usage{
ProcessMatlabDTree(edge_table, predecessors)
}
\arguments{
\item{edge_table}{a numeric matrix whose rows are undirected edges of a tree,
col 1 is v1, col2 is v2, col3 is weight}

\item{predecessors}{a vector of tree predecessors such that pred[i] = the predecessor of i}
}
\value{
a weighted, directed edge table
}
\description{
Convert a matlab edge table and predecessor list into a directed weighted edge table
}
