% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{PseudotimeScatterPlot}
\alias{PseudotimeScatterPlot}
\title{Produce a scatter plot of the cells on selected 2-dim ebedding colored by pseudotime
Here pseudotime is defined as the distance from the root cell according to
the pseudotime metric recorded in \code{pseudotime}}
\usage{
PseudotimeScatterPlot(flat_embedding, pseudotime, outputdir = NULL,
  outputfile = NULL)
}
\arguments{
\item{flat_embedding}{a low dim embedding of cells}

\item{pseudotime}{a scalar representation of pseudotime}

\item{outputdir}{the output directory, relative to getwd()}

\item{outputfile}{the output file}
}
\value{
a ggplot2 object
}
\description{
If an output dir and filename are provided, a plot will
be saved, otherwise just return the plot
}
