% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadAnnotateCountsSCE.R
\name{ReadAnnotateCountsSCE}
\alias{ReadAnnotateCountsSCE}
\title{Read Counts Data}
\usage{
ReadAnnotateCountsSCE(filepath = NULL, inf_exp_rate = 0,
  inf_gene_count = 0, inf_cell_count = 0, series_accession,
  metad_title = "characteristics_ch1", cell_id = "title",
  pheno_title = "Phenotype", pheno_filter = NULL, plot_title = NULL,
  GEOfilecache = getwd(), ...)
}
\arguments{
\item{filepath}{the full or relative (to the workspace) path of the data file}

\item{inf_exp_rate}{the minimum mean expression rate of retained genes}

\item{inf_gene_count}{retain a gene if more than this number of cells express it}

\item{inf_cell_count}{retain a cell if more than this number of genes are expressed in it}

\item{series_accession}{the series accession code, eg "GSE67602"}

\item{metad_title}{the series accession code, eg "GSE67602"}

\item{cell_id}{the series accession code, eg "GSE67602"}

\item{pheno_title}{the name of the phenotype to add to sce, gotten from metad_title}

\item{pheno_filter}{a vector of strings to partially match cell metadata}

\item{plot_title}{path of the saved mean/variance plot, default NULL, if NULL the plot will be printed instead}

\item{GEOfilecache}{location for caching series supplemental files, default is current directory}

\item{...}{arguments to called functions}
}
\value{
a list containing:
    \item{c(seur, sce)}{a seurat or sce object}
}
\description{
Read counts data from file and annotate the generated Single Cell Experiment Object.  Right now this function assumes:
- mouse cells
- only spike ins are ERCC
- row and column names are provided in the data
- the row names are gene symbols
In the case of GEO data, it takes a GSE accession code and assumes that the supp files contain all the expression data
}
