% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadFilter10X.R
\name{ReadFilter10X}
\alias{ReadFilter10X}
\title{Read 10X Data}
\usage{
ReadFilter10X(filepath = NULL, inf_exp_rate = 0, inf_gene_count = 0,
  inf_cell_count = 0, annotation_db, feature_annotation_key,
  remove_spike_ERCC = TRUE, pre_clust = FALSE, rescale = TRUE,
  subsample = NULL, filtercells = TRUE, var_gene_method = "mean_var",
  ...)
}
\arguments{
\item{filepath}{full path of the data directory}

\item{inf_exp_rate}{the minimum mean expression rate of retained genes}

\item{inf_gene_count}{retain a gene if more than this number of cells express it}

\item{inf_cell_count}{retain a cell if more than this number of genes are expressed in it}

\item{annotation_db}{annotation package for the data, eg "org.Mm.eg.db"}

\item{feature_annotation_key}{the key for annotation.db lookups, eg "SYMBOL"}

\item{remove_spike_ERCC}{remove ERCC spike ins after they are used for normalization}

\item{pre_clust}{whether or not to precluster the cells before getting size factors}

\item{rescale}{whether to compute size factors and rescale counts}

\item{subsample}{randomly pick n cells}

\item{filtercells}{whether to do QC in scater}

\item{var_gene_method}{how to select variable genes, eg 'mean_var', 'pca'.  mean_var selects the genes whose variance is highest after removing the fitted spike-in-based technical variance. pca finds the genes with the max rotation coefficients}

\item{...}{arguments to called functions}
}
\value{
a list containing:
    \item{c(seur, sce)}{a seurat or sce object}
}
\description{
Read data from 10X counts or per-molecule info.
}
