% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepresentationMap.R
\name{RepresentationMap}
\alias{RepresentationMap}
\title{Generate multiple representations}
\usage{
RepresentationMap(
  flat_embedding = NULL,
  similarity_matrix,
  join_components = TRUE,
  normalize_S = TRUE,
  flat_embedding_method = "umap",
  ...
)
}
\arguments{
\item{flat_embedding}{optionally provided low dim embedding, if not then 2d tsne will be used}

\item{similarity_matrix}{the graphical embedding of the cells}

\item{join_components}{boolean, whether or not to join disconnected components of the similarity matrix}

\item{normalize_S}{whether or not to normalize the similarity matrix}

\item{flat_embedding_method}{what method to use for flat embedding.  Use umap by default.  Can be umap or tsne.}

\item{...}{arguments to called functions}
}
\value{
a list containing:
    \item{dist_flat}{distance matrix on flat embedding}
    \item{dist_graph}{distance matrix on graph}
    \item{adj_matrix}{unweighted adjacency matrix}
    \item{flat_embedding}{the low dimensional embedding of the cells}
    \item{similarity_graph}{igraph object on the unweighted adjacency matrix}
    \item{components}{an igraph components object based on \code{similarity_graph}}
    \item{n_components}{number of components in \code{components}}
    \item{sizes}{sizes of components in \code{components}}
    \item{members}{list of members of \code{components}, sorted by component size and member index}
}
\description{
Generate representations of the data to be used in lineage analysis.
}
