% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSelection.R
\name{ScaleCenterData}
\alias{ScaleCenterData}
\title{Scale and center a data matrix}
\usage{
ScaleCenterData(M)
}
\arguments{
\item{M}{a matrix of expression values for each gene (rows) and cell (columns)}
}
\value{
a scaled and centered matrix of expression values
}
\description{
We want a matrix whose values are standard deviations from the mean and which are centered at zero.
}
