% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{SigPlot_Cluster}
\alias{SigPlot_Cluster}
\title{Heatmap plot of cluster signaling}
\usage{
SigPlot_Cluster(
  p,
  labels,
  textsize = 30,
  plottitle = NULL,
  plotsubtitle = NULL
)
}
\arguments{
\item{p}{signaling probabilities clusters x clusters}

\item{labels}{labels of cells ordered from 1 to n}

\item{textsize}{the ggplot text size}

\item{plottitle}{the plot title}

\item{plotsubtitle}{the plot sub title}
}
\description{
Produce a plot of a cluster signaling network according to the cell-normalized expression score.
}
