% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimilarityM.R
\name{SimilarityM}
\alias{SimilarityM}
\title{Compute the similarity matrix}
\usage{
SimilarityM(lambda, data)
}
\arguments{
\item{lambda}{the balance term between the rank of Z and the error}

\item{data}{the expression data, where each column is treated as a normalized vector}
}
\value{
a list containing the symetric cell to cell similarity matrix and
    manifold learning error
}
\description{
Computes low dim embedding, constructs KNN graph on the embedding -> unweighted adjacency
Calls manifold learning algorithm which uses the normalized sample vectors and the
unweighted adjacency matrix to compute a low rank approximation of the data
}
