% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signaling.R
\name{SimpleCap}
\alias{SimpleCap}
\title{Capitalize the first letter of a string}
\usage{
SimpleCap(x)
}
\arguments{
\item{x}{the string to run}
}
\value{
a string
}
\description{
Capitalize the first letter of a string.
}
