% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SymNMF.R
\name{SymNMF}
\alias{SymNMF}
\title{Factor a matrix based on an initial factor guess from SVD}
\usage{
SymNMF(A, nC, H, gamma = 1e-06, mu = 10^(-6), maxiter = 1e+06)
}
\arguments{
\item{A}{The similarity matrix.}

\item{k}{The rank of the output.}
}
\value{
\code{W} and \code{H} such that \code{A} = \code{W} * \code{D} * \code{H}.
}
\description{
Factor a matrix based on an initial factor guess from SVD
}
