% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signaling.R
\name{updateFlatDown}
\alias{updateFlatDown}
\title{compute the unnormalized probability}
\usage{
updateFlatDown(ligname, recname, gamma, flat, data)
}
\arguments{
\item{ligname}{the ligand}

\item{recname}{the receptor}

\item{gamma}{the down targets}

\item{flat}{the flattened p matrix}

\item{data}{the expression}
}
\value{
a vector of probs
}
\description{
compute the unnormalized probability
}
