% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signaling.R
\name{updateFlatNone}
\alias{updateFlatNone}
\title{compute the unnormalized probability}
\usage{
updateFlatNone(ligname, recname, flat, data)
}
\arguments{
\item{ligname}{the ligand}

\item{recname}{the receptor}

\item{flat}{the flattened p matrix}

\item{data}{the expression}
}
\value{
a vector of probs
}
\description{
compute the unnormalized probability
}
