% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signaling.R
\name{updateFlatUp}
\alias{updateFlatUp}
\title{compute the unnormalized probability}
\usage{
updateFlatUp(ligname, recname, beta, flat, data)
}
\arguments{
\item{ligname}{the ligand}

\item{recname}{the receptor}

\item{beta}{the up targets}

\item{flat}{the flattened p matrix}

\item{data}{the expression}
}
\value{
a vector of probs
}
\description{
compute the unnormalized probability
}
