% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{ViolinPlotExpression}
\alias{ViolinPlotExpression}
\title{Violin plot of gene expression}
\usage{
ViolinPlotExpression(data, gene_names, labels, gene_name,
  colorscale = NULL)
}
\arguments{
\item{data}{a matrix with genes in rows and cells in columns}

\item{gene_names}{a vector of names corresponding to the rows in data}

\item{labels}{a vector of cluster assignments for the cells}

\item{gene_name}{the name of the gene to plot}

\item{colorscale}{optionally provided color scale}
}
\value{
nothing
}
\description{
Produce a violin plot of gene expression.
}
