% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{ChordColors}
\alias{ChordColors}
\title{Get chord colors for cells}
\usage{
ChordColors(edge_table, cluster_cols, gap)
}
\arguments{
\item{edge_table}{a table with lig, rec, score, lig_cluster, rec_cluster}

\item{cluster_cols}{colors for the clusters}

\item{gap}{the hue gap}
}
\description{
Get chord colors for cells
}
