% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signaling.R
\name{ClusterSig}
\alias{ClusterSig}
\title{Average cluster to cluster signaling}
\usage{
ClusterSig(P, cluster_labels)
}
\arguments{
\item{P}{signaling probabilities cells x cells}

\item{cluster_labels}{labels of cells 1:n}
}
\value{
a matrix of cluster to cluster signaling
}
\description{
Average cluster to cluster signaling.
}
