% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{FeatureScatterPlot}
\alias{FeatureScatterPlot}
\title{Produce a scatter plot of the cells}
\usage{
FeatureScatterPlot(
  flat_embedding,
  feature,
  title,
  subtitle,
  featurename,
  colorscale = NULL,
  pointsize = 1,
  hide_legend = FALSE,
  legsize = 5,
  legt = 5,
  legtxt = 5,
  grad_lim = NULL,
  autoscale_dropout = FALSE,
  dropout_NA = TRUE
)
}
\arguments{
\item{flat_embedding}{a low dim embedding of cells}

\item{feature}{a scalar representation of feature}

\item{title}{the title of the plot}

\item{subtitle}{the subtitle of the plot}

\item{featurename}{the name of the feature to title the legend}

\item{colorscale}{color scale for discrete data}

\item{pointsize}{the size of geom point}

\item{hide_legend}{hide the legend and return a list, containing the legendless plot and the legend}

\item{legsize}{the size of the legend}

\item{legt}{size of legend title}

\item{legtxt}{size of legend text}

\item{grad_lim}{the limits of the color gradient for continuous features, this should usually be in log-expression space, overrides autoscale_dropout}

\item{autoscale_dropout}{whether to set plots where the feature space is 0 to #102033, default min for scale_color_gradient auto scaling, overrides dropout_NA}

\item{dropout_NA}{whether to set plots where the feature space is 0 to NA, note: the legend element in a hide_legend output will be NULL}
}
\value{
a ggplot2 object
}
\description{
Create a scatter plot of the data on a 2-dim ebedding colored by feature.  If discrete (factor) data is passed, the colorscale parameter must be provided, otherwise the plot will default to gradient shading.
}
