% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{FeatureScatterPlot}
\alias{FeatureScatterPlot}
\title{Produce a scatter plot of the cells}
\usage{
FeatureScatterPlot(flat_embedding, feature, title, subtitle, featurename,
  colorscale = NULL)
}
\arguments{
\item{flat_embedding}{a low dim embedding of cells}

\item{feature}{a scalar representation of feature}

\item{title}{the title of the plot}

\item{subtitle}{the subtitle of the plot}

\item{featurename}{the name of the feature to title the legend}

\item{colorscale}{color scale for discrete data}
}
\value{
a ggplot2 object
}
\description{
Create a scatter plot of the data on a 2-dim ebedding colored by feature.  If discrete (factor) data is passed, the colorscale parameter must be provided, otherwise the plot will default to gradient shading.
}
