% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindRootCell.R
\name{FindRootCell}
\alias{FindRootCell}
\title{Find the root cell}
\usage{
FindRootCell(use_flat_dist = TRUE, cluster_order_by = "distance",
  cell_order_by = "distance", graph_cluster_mst = NULL,
  dist_graph = NULL, dist_flat, cluster_labels = NULL,
  root_cluster = NULL)
}
\arguments{
\item{use_flat_dist}{use max path length heuristic on flat embedding}

\item{cluster_order_by}{cluster rank parameter for measuring cluster/cell rank correlation}

\item{cell_order_by}{cell-wise pseudotime parameter for measuring cluster/cell rank correlation}

\item{graph_cluster_mst}{igraph object representing an mst on the cluster-cluster graph}

\item{dist_graph}{a distance matrix of cells embedded in a graph}

\item{dist_flat}{the manifold embedding of the cells}

\item{cluster_labels}{the cluster label for each cell}

\item{root_cluster}{the id of the root cluster on the cluster-cluster graph}
}
\value{
integer index of the root cell
}
\description{
Use either the primary manifold embedding of cell similarity to find the root cell by cluster-cell rank-correlation
or use the flattened representation of this embedding to find the root cell by maximum separation heuristic
possible values for cluster_order_by are predecessor and distance
possible values for cell_order_by are index and distance
}
