% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterNumber.R
\name{GetConsensus}
\alias{GetConsensus}
\title{Produce a consensus matrix}
\usage{
GetConsensus(clusters)
}
\arguments{
\item{clusters}{the cluster assignment of each cell}
}
\value{
a consensus matrix
}
\description{
Produce a consensus matrix for ensemble analysis.
}
