% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeFunctions.R
\name{GetDominatorTree}
\alias{GetDominatorTree}
\title{Get a directed graph from a predecessor vector}
\usage{
GetDominatorTree(predecessors, weighted_graph = NULL)
}
\arguments{
\item{predecessors}{a predecessor vector, such that predecessors[i] = the predecessor of i}

\item{weighted_graph}{a weighted igraph object}
}
\value{
a directed igraph object
}
\description{
Get a directed graph from a predecessor vector
}
