% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkerGenes.R
\name{GetMarkerTable}
\alias{GetMarkerTable}
\title{Get the marker genes for each cluster}
\usage{
GetMarkerTable(counts_data, cluster_labels, H, gene_expression_threshold,
  n_features)
}
\arguments{
\item{cluster_labels}{a vector of cluster labels}

\item{H}{a nonnegative matrix such that W = H*t(H), H_{i,j} is the cells_weight
by which cell i belongs to the jth cluster}

\item{gene_expression_threshold}{for n cells, for \code{gene_expression_threshold} = m, dont consider genes
expressed in more than n-m cells or genes expressed in less than m cells}

\item{n_features}{number of marker genes per cluster to retrieve}

\item{M}{a matrix of expression values for each cell (rows) and gene (columns)}
}
\value{
a table of marker genes
}
\description{
Get the marker genes for each cluster
}
