% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signaling.R
\name{GetSignalingPartners}
\alias{GetSignalingPartners}
\title{Compute Cell-cell interaction probability}
\usage{
GetSignalingPartners(M, ids, pathway, normalize_aggregate = TRUE)
}
\arguments{
\item{M}{a matrix of expression values for each cell (rows) and gene (columns)}

\item{ids}{a vector of gene ids}

\item{pathway}{a data frame with "ligands", "receptors", "direction", and "targets"}

\item{normalize_aggregate}{whether or not to normalize the P_agg matrix by dividing each row by its sum default is true}
}
\value{
a list containing:
    \item{P}{a list of the cell-cell signaling probabilities for all ligand/receptor pairs}
    \item{P_agg}{the aggregate matrix of all ligand/receptor pairs}
}
\description{
We can have a situation where ligand/receptor expression is low, but target genes (repressors and activators) are highly expressed.  In this case we will get a false positive for \eqn{P_{i,j}}.  In order to correct for this, we introduce normalizing coefficients for the relationship between \eqn{\alpha_{i,j}} and \eqn{\beta / \gamma}.  When \eqn{\alpha} is low, then K and D will decrease rapidly with increasing \eqn{\beta} and \eqn{\gamma}, penalizing the resulting increase in P (which is increasingly likely to be a false positive, given the disparity between \eqn{\alpha} and \eqn{\beta},\eqn{\gamma})
\eqn{K_{i,j} = \frac{\alpha_{i,j}}{\alpha_{i,j} + \beta_{i,j}}}
Note that activated genes are required under this model.
}
