% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signaling.R
\name{PenaltyCoeff}
\alias{PenaltyCoeff}
\title{Generate the divergence penalty coefficient}
\usage{
PenaltyCoeff(alpha, const, num = "alpha", n_cells)
}
\arguments{
\item{alpha}{a matrix of normalized LR expression values for each cell (columns) and gene (rows)}

\item{const}{either the normalized repression targets or the normalized expression targets}

\item{num}{either alpha or const}

\item{n_cells}{the number of cells (row-traversal of the alpha matrix)}
}
\value{
the K or D penalty for a signaling matrix
}
\description{
This will approach 1 as the difference between alpha and \code{const} approaches 0.
}
