% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{PlotClusterExpression}
\alias{PlotClusterExpression}
\title{Heatmap of specific genes}
\usage{
PlotClusterExpression(data, gene_names, cluster_labels, markers)
}
\arguments{
\item{data}{expression data with genes x cells}

\item{gene_names}{a vector of symbolic gene names corresponding to the rows in the data matrix}

\item{cluster_labels}{a vector of cluster labels corresponding to the columns in the data matrix}

\item{markers}{a vector of markers}
}
\value{
nothing
}
\description{
Given a set of markers, plot their expression across clusters on a heatmap.
}
