% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{PlotLineage}
\alias{PlotLineage}
\title{Plot the pseudotime ordering of clusters}
\usage{
PlotLineage(network, d_thickness = c(1, 2), node_color = NULL,
  alpha_color = 1)
}
\arguments{
\item{network}{an igraph network with weighted edges corresponding to the pseudotime distance}

\item{d_thickness}{controls edge width.  c(1, m) => m/distance from L2R2, c(0,m) => m.}

\item{node_color}{an optional vector of colors for node labeling.  If null then the nodes are all colored black}

\item{alpha_color}{the gradient of the nodes}
}
\value{
nothing
}
\description{
Plot the pseudotime ordering of clusters
}
