% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{PlotLineage}
\alias{PlotLineage}
\title{Plot the pseudotime ordering of clusters}
\usage{
PlotLineage(network, node_color = NULL, alpha_color = 1,
  exaggerate = TRUE, weight_scaler = 0.2, arrow_scaler = 0.8)
}
\arguments{
\item{network}{an igraph network with weighted edges corresponding to the pseudotime distance}

\item{node_color}{an optional vector of colors for node labeling.  If null then the nodes are all colored black}

\item{alpha_color}{the gradient of the nodes}

\item{exaggerate}{whether to exponentiate then linearly scale the edge weights, thus emphasizing their differences}

\item{weight_scaler}{the weight scaler for edges}

\item{arrow_scaler}{the arrow width scaler for edges}
}
\value{
nothing
}
\description{
Plot the pseudotime ordering of clusters
}
