% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{PlotTopN}
\alias{PlotTopN}
\title{Heatmap of the top n markers}
\usage{
PlotTopN(data, gene_names, cluster_labels, markers, n_features)
}
\arguments{
\item{data}{expression data with genes x cells}

\item{gene_names}{a vector of symbolic gene names corresponding to the rows in the data matrix}

\item{cluster_labels}{a vector of cluster labels corresponding to the columns in the data matrix}

\item{markers}{a table of markers}

\item{n_features}{the top n features to plot}
}
\value{
nothing
}
\description{
Plot the heatmap of the top n inferred markers for each cluster.
}
