% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepresentationMap.R
\name{RepresentationMap}
\alias{RepresentationMap}
\title{generate convenient representations of the data
representations of the data are necessary for subsequent lineage analysis}
\usage{
RepresentationMap(flat_embedding = NULL, similarity_matrix,
  join_components = TRUE, ...)
}
\arguments{
\item{flat_embedding}{optionally provided low dim embedding, if not then 2d tsne will be used}

\item{similarity_matrix}{the graphical embedding of the cells}

\item{join_components}{boolean, whether or not to join disconnected components of the similarity matrix}

\item{...}{arguments to called functions}
}
\value{
a list containing:
    \item{dist_flat}{distance matrix on flat embedding}
    \item{dist_graph}{distance matrix on graph}
    \item{adj_matrix}{unweighted adjacency matrix}
    \item{flat_embedding}{the low dimensional embedding of the cells}
    \item{similarity_graph}{igraph object on the unweighted adjacency matrix}
    \item{components}{an igraph components object based on \code{similarity_graph}}
    \item{n_components}{number of components in \code{components}}
    \item{sizes}{sizes of components in \code{components}}
    \item{members}{list of members of \code{components}, sorted by component size and member index}
}
\description{
generate convenient representations of the data
representations of the data are necessary for subsequent lineage analysis
}
\examples{
embeddings <- RepresentationMap(flat_embedding = RSoptSC::GuoPtime$Params$latent,
    similarity_matrix = RSoptSC::GuoPtime$Params$W,
    join_components = FALSE)

}
