% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{SigPlot}
\alias{SigPlot}
\title{Heatmap plot of cell signaling}
\usage{
SigPlot(p, labels, textsize = 10, plottitle = NULL, plotsubtitle = NULL)
}
\arguments{
\item{p}{signaling probabilities cells x cells}

\item{labels}{labels of cells ordered from 1 to n}

\item{textsize}{the ggplot text size}

\item{plottitle}{the plot title}

\item{plotsubtitle}{the plot sub title}
}
\description{
Produce a plot of a cell signaling network according to the cell-normalized expression score.
}
