% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{SigPlot}
\alias{SigPlot}
\title{Circos plot of signaling score}
\usage{
SigPlot(P, rgb_gap = 0.2, cluster_labels, lig_cells_per_cluster = 10,
  lig_cells_total = 0, lig_edge_per_cell = 0,
  lig_edge_per_cluster = 0, lig_edge_total = 0,
  rec_cells_per_cluster = 10, rec_cells_total = 0,
  zero_threshold = 0)
}
\arguments{
\item{P}{signaling probabilities cells x cells}

\item{rgb_gap}{real percent of the inter-cluster spectrum gap to encompass cluster highlighting}

\item{cluster_labels}{labels of cells ordered from 1 to n}

\item{lig_cells_per_cluster}{int 0 to keep all}

\item{lig_cells_total}{int 0 to keep remaining cells per cluster}

\item{lig_edge_per_cell}{int 0 to keep all}

\item{lig_edge_per_cluster}{int 0 to keep remaining edge per cell}

\item{lig_edge_total}{int 0 to keep keep remaining edge per cluster}

\item{rec_cells_per_cluster}{int 0 to keep all}

\item{rec_cells_total}{int 0 to keep remaining cells per cluster}

\item{zero_threshold}{real value for zero cutoff (0 to keep all edges)}
}
\description{
Produce a plot of a cell signaling network according to the cell-normalized expression score.
}
