% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SymNMF-old.R, R/SymNMF.R
\name{SymNMF}
\alias{SymNMF}
\title{Initialize non-negative factorization of the similarity matrix}
\usage{
SymNMF(A, nC, H, gamma = 1e-06, mu = 10^(-6), maxiter = 1e+06)

SymNMF(A, nC, H, gamma = 1e-06, mu = 10^(-6), maxiter = 1e+06)
}
\arguments{
\item{A}{The similarity matrix.}

\item{k}{The rank of the output.}

\item{A}{The similarity matrix.}

\item{k}{The rank of the output.}
}
\value{
\code{W} and \code{H} such that \code{A} = \code{W} * \code{D} * \code{H}.

\code{W} and \code{H} such that \code{A} = \code{W} * \code{D} * \code{H}.
}
\description{
Initialize non-negative factorization of the similarity matrix

Initialize non-negative factorization of the similarity matrix
}
\examples{
InitSVD(A, k)
InitSVD(A, k)
}
