% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signaling.R
\name{TargetAvg}
\alias{TargetAvg}
\title{Get average target expression for each cell}
\usage{
TargetAvg(M, ids)
}
\arguments{
\item{M}{a matrix of expression values for each cell (rows) and gene (columns)}

\item{ind}{a vector of row indexes corresponding to genes}
}
\description{
Get average target expression for each cell
}
